% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARKS_analysis.R
\name{add_custom_library_to_SPARKS_test_result}
\alias{add_custom_library_to_SPARKS_test_result}
\title{Add SPAKRS result for single AS signature}
\usage{
add_custom_library_to_SPARKS_test_result(
  input_mats,
  input_result,
  test_mats,
  test_study,
  kd_library = list(),
  event_count_df = data.frame(),
  library_size = 675
)
}
\arguments{
\item{input_mats}{MATS df for the study of interest}

\item{input_result}{SPARKS analysis result against AS signature library}

\item{test_mats}{MATS df for the signature}

\item{test_study}{signature name}

\item{kd_library}{If given, this entire library will be used to compute the AS events fo filter out}

\item{event_count_df}{If given, this list will be used to compute the AS events fo filter out. This is more memory-efficient}

\item{library_size}{This is 675 for now. Don't change}
}
\value{
SPARKS analysis result with new result added
}
\description{
Add SPAKRS result for single AS signature
}
