% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lollipop_summary_plot.R
\name{add_plot_title}
\alias{add_plot_title}
\title{Add title to combined plot}
\usage{
add_plot_title(
  p_combined,
  plot_title,
  title_ratio = 10,
  left_margin = 0,
  title_color = "black"
)
}
\description{
Add title to combined plot
}
