% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_splice_site.R
\name{calculate_motif_occurrence}
\alias{calculate_motif_occurrence}
\title{Query Splice site sequence for given AS events}
\usage{
calculate_motif_occurrence(as_events, padding = 100, motif = "CCCCC")
}
\arguments{
\item{as_events}{AS events to get query for}
}
\value{
splice site information object sith ss sequence and ss score
}
\description{
Query Splice site sequence for given AS events
}
