% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARKS_analysis.R
\name{calculate_new_psi_for_subset_of_samples}
\alias{calculate_new_psi_for_subset_of_samples}
\title{Calculate new MATS df for subset of samples}
\usage{
calculate_new_psi_for_subset_of_samples(
  study_mats,
  comp_sample_id_1,
  comp_sample_id_2,
  count_threshold = 20,
  min_threshold = 20
)
}
\arguments{
\item{study_mats}{Filtered MATS df}

\item{comp_sample_id_1}{Column ID for group one.}

\item{comp_sample_id_2}{Column ID for group two}

\item{count_threshold}{Average read count for AS events. Events with read counts below this number in average is removed}

\item{min_threshold}{Minimum read count for AS events. Events with read counts below this number in any sample is removed}
}
\value{
subset_mats_df Filtered MATS df for the subset of the samples
}
\description{
The sample column ID can be identified by PSI df stored in SPARKS object
}
