% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_hotspot_association.R
\name{calculate_sig_AS_event_per_HS}
\alias{calculate_sig_AS_event_per_HS}
\title{Calculate number of associated AS events for each hotpsot in event type}
\usage{
calculate_sig_AS_event_per_HS(input_lm_df, event_type)
}
\arguments{
\item{input_lm_df}{input LM df from the pipeline}

\item{event_type}{SE, A3SS or A5SS}
}
\value{
count df
}
\description{
Calculate number of associated AS events for each hotpsot in event type
}
