% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARKS_analysis.R
\name{calculate_SPARKS_padj}
\alias{calculate_SPARKS_padj}
\title{Calculate adjusted p-value for SPARKS analysis results}
\usage{
calculate_SPARKS_padj(
  sparks_result,
  sig_test_method = "bonferroni",
  same_sign_ES_threshold = 0.5
)
}
\arguments{
\item{sparks_result}{SPARKS analysis dataframe}

\item{sig_test_method}{p.adjust method - "BH" for FDR, "bonferroni" (default) for FWER}

\item{same_sign_ES_threshold}{Set any signatures with abs score below this when sign from pos and neg signature sets are the same}
}
\value{
SPARKS analysis dataframe with padj added
}
\description{
This function sets the padj.
Any signatures with the same sign between positive ES and nevative ES are set
to have padj of 1, so they wouldn't be considered significant.
This approach is similar to C-MAP paper (Lamb et al., Science 2006) that
they remove any signatures with same sign are removed.
}
