% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_CLIP_data.R
\name{call_CLIP_targets_for_ENCODE_data}
\alias{call_CLIP_targets_for_ENCODE_data}
\title{Generate CLIP peak annotated DS events}
\usage{
call_CLIP_targets_for_ENCODE_data(object, target, event_type = "SE")
}
\arguments{
\item{object}{SPARKS object}

\item{target}{Target RBP}

\item{event_type}{either SE, A3SS, A5SS}
}
\value{
result dataframe with clip peak presence and beta from MATS
}
\description{
Generate CLIP peak annotated DS events
}
