% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{classify_cosmic_genes}
\alias{classify_cosmic_genes}
\title{Parse COSMIC data}
\usage{
classify_cosmic_genes(cosmic_df)
}
\arguments{
\item{cosmoc_df}{COSMIC df read in from COSMIC data from website}
}
\value{
list of genes grouped in TSG/OG/None separately
}
\description{
Parses COSMIC genes to classify either Tumor suppressing or oncogene
}
