% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_hotspot_association.R
\name{classify_hotspot_mutation}
\alias{classify_hotspot_mutation}
\title{Classify hotspot mutations}
\usage{
classify_hotspot_mutation(plot_df)
}
\arguments{
\item{plot_df}{hotspot plot df}
}
\value{
mutation class df determined for each hotspot
}
\description{
Classify hotspot mutations
}
