% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{cleanup_exp_df}
\alias{cleanup_exp_df}
\title{Remove duplicates arising from sample processing}
\usage{
cleanup_exp_df(psi_df)
}
\arguments{
\item{exp_df}{EXP value dataframe}
}
\value{
clean_exp_df - PSI df without duplicated samples
}
\description{
This function removes columns (samples) in EXP df where there may be duplicates
generated from some error in processing pipeline
}
