% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{cleanup_hotspot_mutation_matrix}
\alias{cleanup_hotspot_mutation_matrix}
\title{Wrapper for cleaning up hotspot mutation type}
\usage{
cleanup_hotspot_mutation_matrix(hs_mut_all)
}
\arguments{
\item{hs_mut_all}{hotspot mutation}
}
\value{
sorted hotspot mutation matrix
}
\description{
This function parses hotspot mutations for downstream visualization
}
