% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{cleanup_hotspot_pval_data}
\alias{cleanup_hotspot_pval_data}
\title{Clean up pval data for downstream analysis}
\usage{
cleanup_hotspot_pval_data(input_pval_df, spl_event_type)
}
\arguments{
\item{input_pval_df}{pval dataframe form analysis script}

\item{spl_event_type}{either SE, A3SS, A5SS}
}
\value{
pval_df where rows are AS events, columns are mutation/sample, elements are -log10(pval)
}
\description{
This function nicely gather and sort hotspot and AS events by coordinates
then convert the information into neglog10 for easier handling
}
