% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{cleanup_psi_df}
\alias{cleanup_psi_df}
\title{Remove duplicates arising from sample processing}
\usage{
cleanup_psi_df(psi_df, event_type)
}
\arguments{
\item{psi_df}{PSI value dataframe}

\item{event_type}{either SE, A3SS, A5SS}
}
\value{
clean_psi_df - PSI df without duplicated samples
}
\description{
This function removes columns (samples) in PSI df where there may be duplicates
generated from some error in processing pipeline
}
