% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rMATS_output.R
\name{convert_MATS_format}
\alias{convert_MATS_format}
\title{convert MATS to SPARKS format}
\usage{
convert_MATS_format(input_mats)
}
\arguments{
\item{input_mats}{filtered ${spl_type}.MATS.JC.txt}
}
\value{
converted MATS
}
\description{
convert MATS to SPARKS format
}
