% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.detect_quant_format}
\alias{.detect_quant_format}
\title{Detect the quantification format of a sample folder}
\usage{
.detect_quant_format(sample_dir)
}
\arguments{
\item{sample_dir}{Character. Path to one sample's folder.}
}
\value{
Character format code, or NA if nothing recognizable was found.
}
\description{
Looks for the on-disk layout of each supported format, in order:
AnnData (.h5ad) > Alevin > Cell Ranger MEX > Cell Ranger HDF5 > STARsolo >
a bare MatrixMarket triplet.
}
\keyword{internal}
