% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.filter_present_genes}
\alias{.filter_present_genes}
\title{Filter requested genes to those present in the object}
\usage{
.filter_present_genes(genes, seurat_obj, context = "")
}
\arguments{
\item{genes}{Character vector}

\item{seurat_obj}{Seurat object}

\item{context}{Character for logging}
}
\value{
Character vector of available genes, or NULL
}
\description{
Filter requested genes to those present in the object
}
