% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.find_all_markers_safe}
\alias{.find_all_markers_safe}
\title{Wrapper for FindAllMarkers to catch errors safely}
\usage{
.find_all_markers_safe(
  seurat_obj,
  only_pos = TRUE,
  min_pct = 0.25,
  logfc_threshold = 0.25
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{only_pos}{Logical}

\item{min_pct}{Numeric}

\item{logfc_threshold}{Numeric}
}
\value{
Data frame of markers or NULL
}
\description{
Wrapper for FindAllMarkers to catch errors safely
}
