% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.has_reduction}
\alias{.has_reduction}
\title{Check if object has a specific reduction}
\usage{
.has_reduction(seurat_obj, reduction)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{reduction}{Character}
}
\value{
Logical
}
\description{
Check if object has a specific reduction
}
