% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.has_scale_data}
\alias{.has_scale_data}
\title{Check if object has scaled data}
\usage{
.has_scale_data(seurat_obj)
}
\arguments{
\item{seurat_obj}{Seurat object}
}
\value{
Logical
}
\description{
Check if object has scaled data
}
