% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.load_generic_mtx}
\alias{.load_generic_mtx}
\title{Read a generic MatrixMarket triplet that doesn't follow 10x file naming}
\usage{
.load_generic_mtx(
  sample_dir,
  matrix_file = "matrix.mtx",
  barcodes_file = "barcodes.tsv",
  features_file = "features.tsv"
)
}
\description{
Read a generic MatrixMarket triplet that doesn't follow 10x file naming
}
\keyword{internal}
