% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.safe_filename}
\alias{.safe_filename}
\title{Sanitize string for filenames}
\usage{
.safe_filename(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Sanitized character vector
}
\description{
Sanitize string for filenames
}
