% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.warn_if_noninteger_counts}
\alias{.warn_if_noninteger_counts}
\title{Warn if a "raw counts" matrix looks like it's actually already normalized}
\usage{
.warn_if_noninteger_counts(cnts, label = "")
}
\description{
Warn if a "raw counts" matrix looks like it's actually already normalized
}
\keyword{internal}
