% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_accessory_functions.R
\name{extract_exp_value_for_AS_event}
\alias{extract_exp_value_for_AS_event}
\title{Query EXP value for AS event}
\usage{
extract_exp_value_for_AS_event(object, AS_event)
}
\arguments{
\item{object}{SPARKS object}

\item{event_type}{AS event}
}
\value{
EXP values for the AS event
}
\description{
Query EXP value for AS event
}
