% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_accessory_functions.R
\name{extract_gene_symbols}
\alias{extract_gene_symbols}
\title{Extract gene symbols from complex annotation}
\usage{
extract_gene_symbols(hotspot_list)
}
\arguments{
\item{hotspot_list}{AS event/hotspot list}
}
\value{
Gene symbol, which is generally in the second place in colon seperated names
}
\description{
Extract gene symbols from complex annotation
}
