% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARKS_analysis.R
\name{extract_GSEA_significant_events}
\alias{extract_GSEA_significant_events}
\title{Extract signature events for SPARKS}
\usage{
extract_GSEA_significant_events(study_mats)
}
\arguments{
\item{study_mats}{filtered MATS df}
}
\value{
event_list List of events with positive (delPSI > 0.1) and negative (delPSI < -0.1) AS changes
}
\description{
This function extracts events with abs(delPSI) > 0.1 and store in a list, sepearated in
positive and negative set of AS events
}
