% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_choosecopy.R
\name{file_choosecopy}
\alias{file_choosecopy}
\title{file_choosecopy}
\usage{
file_choosecopy(x)
}
\arguments{
\item{hi}{}
}
\value{
hi
}
\description{
addition to file_choose: it copies the path into clipboard. only works on linux.
}
\examples{
hi

}
