% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_outliers.R
\name{find_outliers}
\alias{find_outliers}
\title{find_outliers}
\usage{
find_outliers(y, coef = 1.5)
}
\arguments{
\item{hi}{}
}
\value{
hi
}
\description{
given a vector, find outliers
}
\examples{
hi

}
