% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findLasso.R
\name{findLasso}
\alias{findLasso}
\title{findLasso}
\usage{
findLasso(dat, zeros, R, whichfunction = c("spca_adj", "rsvd_spca"),
  init, ridge = 1e-06, maxiterOut, maxiterIn)
}
\arguments{
\item{dat}{dataset}

\item{zeros}{vector: number of zeros wanted per component}

\item{R}{number of components}

\item{whichfunction}{choice between "spca_adj" or "rsvd_spca": which function do you want?}

\item{init}{initial values for the lasso penalties}

\item{ridge}{ridge penalty. default 1e-6}

\item{maxiterOut}{maximum iteration for all}

\item{maxiterIn}{maximum iteration within one component}
}
\value{
lasso, converged
}
\description{
major idea/code originally by Niek. Thanks sir. Provides the lasso penalty values, given the number of zeros in the PCA coefficients matrix
}
\examples{
hi

}
