% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_CLIP_data.R
\name{generate_CLIP_enrichment_heatmap}
\alias{generate_CLIP_enrichment_heatmap}
\title{Generate CLIP Peak enrichment HEATMAP}
\usage{
generate_CLIP_enrichment_heatmap(region_pval_select, event_type = "SE")
}
\arguments{
\item{region_pval_select}{result df from CLIP peak enrichment test}

\item{event_type}{either SE, A3SS, A5SS}
}
\value{
heatmap generated by ggplot
}
\description{
Generate CLIP Peak enrichment HEATMAP
}
