% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_cluster_markers_and_heatmap}
\alias{generate_cluster_markers_and_heatmap}
\title{FindAllMarkers heatmap for one grouping}
\usage{
generate_cluster_markers_and_heatmap(
  seurat_obj,
  group_by_col,
  out_dir,
  prefix,
  top_n = 5L,
  min_pct = 0.25,
  logfc_threshold = 0.25
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{group_by_col}{Character}

\item{out_dir}{Character}

\item{prefix}{Character}

\item{top_n}{Integer. Top markers per cluster}

\item{min_pct}{Numeric}

\item{logfc_threshold}{Numeric}
}
\description{
FindAllMarkers heatmap for one grouping
}
