% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_splice_site.R
\name{generate_combined_splice_site_plot}
\alias{generate_combined_splice_site_plot}
\title{Combining script to generate Splice Site Analysis Plot}
\usage{
generate_combined_splice_site_plot(
  p_list_inc,
  p_list_exc,
  p_list_score,
  condition_label_one,
  condition_label_two
)
}
\arguments{
\item{p_list_inc}{list of plots for positive delta psi}

\item{p_list_exc}{list of plots for negative delta psi}

\item{p_list_score}{list of plots for splice site score distribution}

\item{condition_label}{labels to be added in the plot}
}
\value{
combined plot with all the plots merged
}
\description{
Combining script to generate Splice Site Analysis Plot
}
