% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_deg_umap}
\alias{generate_deg_umap}
\title{UMAP coloured by number of DEGs per cluster}
\usage{
generate_deg_umap(
  seurat_obj,
  deg_counts,
  umap_dir,
  file_prefix,
  group_by_col,
  color_high = "red",
  color_low = "lightgrey",
  min_deg_display = 5L,
  reduction = "umap",
  point_size = 0.5,
  alpha = 0.6,
  label_size = 4,
  width = 8,
  height = 7
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{deg_counts}{data.frame with columns Group and DEG_Count, or NULL}

\item{umap_dir}{Character}

\item{file_prefix}{Character}

\item{group_by_col}{Character}

\item{color_high}{Character}

\item{color_low}{Character}

\item{min_deg_display}{Integer. Min DEG_Count to label a cluster}

\item{reduction}{Character}

\item{point_size}{Numeric}

\item{alpha}{Numeric}

\item{label_size}{Numeric}

\item{width}{Numeric}

\item{height}{Numeric}
}
\description{
UMAP coloured by number of DEGs per cluster
}
