% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_barplot.R
\name{generate_enrichment_barplot}
\alias{generate_enrichment_barplot}
\title{Generate enrichment bar plot for SPARKS result}
\usage{
generate_enrichment_barplot(
  result_df,
  bar_color = "grey75",
  sig_threshold = 0.01,
  rank_label = T,
  sig_test_method = "bonferroni",
  num_plot = NA,
  select_genes = c(),
  manual_colors = list(),
  text_scale_factor = 1,
  select_gene_marker = F
)
}
\arguments{
\item{result_df}{SPARKS analysis result}

\item{bar_color}{Color for the bars - default = grey75}

\item{sig_threshold}{FWER threshold - default = 0.01}

\item{rank_label}{If true (default), show ranking of each signatures on top}

\item{sig_test_method}{Multiple testing correction method - default = "bonferroni"}

\item{num_plot}{Number of signatures to plot, top N signatures based on abs(ES) are shown}

\item{select_genes}{If provided, signatures perturbing these genes would be shown in the plot, regardless of their rank/ES. This function also generates rank and FWER indicator on top of the bar}

\item{manual_colors}{List of genes and color for their text e.g. list('SRSF1' = 'red') will show any signatures with SRSF1 perturbation in red}

\item{text_scale_factor}{Scale factor for shiny app}

\item{select_gene_marker}{Explanation for rank and FWER indicator on top of the bar - Use if select genes are activated and don't want to add additional label for the indicators}
}
\value{
Enrichment bar plot
}
\description{
Generate enrichment bar plot for SPARKS result
}
