% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escape_analysis.R
\name{generate_escape_plots}
\alias{generate_escape_plots}
\title{Generate and save escape heatmaps}
\usage{
generate_escape_plots(
  seurat_obj,
  method = "ssGSEA",
  group_col,
  out_dir,
  prefix
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{method}{Character}

\item{group_col}{Character}

\item{out_dir}{Character}

\item{prefix}{Character}
}
\description{
Generate and save escape heatmaps
}
