% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_hotspot_association.R
\name{generate_EXP_boxplot_for_mutation_group}
\alias{generate_EXP_boxplot_for_mutation_group}
\title{Generate boxplot for EExpression depending on hotspot mutation status}
\usage{
generate_EXP_boxplot_for_mutation_group(
  object,
  hotspot_event,
  AS_event,
  exp_df = exp_df_clean,
  hotspot_status = hotspot_status_df
)
}
\arguments{
\item{object}{SPARKS object}

\item{hotspot_event}{Mutation event}

\item{AS_event}{AS event - the gene would be selected by gene symbol}
}
\value{
boxplot with sample grouped by mutation status
}
\description{
This function generates heatmap plot - x = hotspot mutation status, y = Expression
}
