% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_feature_plots}
\alias{generate_feature_plots}
\title{Save a FeaturePlot for each gene in the list}
\usage{
generate_feature_plots(seurat_obj, genes, out_dir, prefix, reduction = "umap")
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{genes}{Character vector}

\item{out_dir}{Character}

\item{prefix}{Character}

\item{reduction}{Character}
}
\description{
Save a FeaturePlot for each gene in the list
}
