% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_hotspot_association.R
\name{generate_hotkey_plot_for_hotspot_mutation_type}
\alias{generate_hotkey_plot_for_hotspot_mutation_type}
\title{Plot function for hot key plot}
\usage{
generate_hotkey_plot_for_hotspot_mutation_type(
  mutation_class_heatmap_df_sorted
)
}
\arguments{
\item{mutation_class_heatmap_df_sorted}{Mutation class heatmap data}
}
\value{
Hotkey plot
}
\description{
This function generates hot key plot - row = hotspot, col = mutation class (e.g. Missense mutation, Nonsense mutation)
}
