% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_hotspot_association.R
\name{generate_hotspot_summary_table_plot}
\alias{generate_hotspot_summary_table_plot}
\title{Wrapper for generating heatmap plot for summary}
\usage{
generate_hotspot_summary_table_plot(
  object,
  genes_of_interest = c(),
  event_group = c()
)
}
\arguments{
\item{object}{SPARKS object}

\item{genes_of_interest}{Genes you are interested in (hotspot)}

\item{event_group}{Group of mutations (e.g. Missense mutation) to limit the plot}
}
\value{
summary heatmap for each mutation and associated AS events
}
\description{
Wrapper for generating heatmap plot for summary
}
