% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_hotspot_association.R
\name{generate_PSI_boxplot_for_mutation_group}
\alias{generate_PSI_boxplot_for_mutation_group}
\title{Generate boxplot for PSI depending on hotspot mutation status}
\usage{
generate_PSI_boxplot_for_mutation_group(object, hotspot_event, AS_event)
}
\arguments{
\item{object}{SPARKS object}

\item{hotspot_event}{Mutation event}

\item{AS_event}{AS event}
}
\value{
boxplot with sample grouped by mutation status
}
\description{
This function generates heatmap plot - x = hotspot mutation status, y = PSI
}
