% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_scatter_plot.R
\name{generate_RBP_KD_correlation_scatter_plot}
\alias{generate_RBP_KD_correlation_scatter_plot}
\title{Generate scatter plot with linear fit statistics between two AS changes}
\usage{
generate_RBP_KD_correlation_scatter_plot(
  s1_mats_df,
  s2_mats_df,
  study_1,
  study_2,
  beta_threshold = 0.1,
  slope = FALSE
)
}
\arguments{
\item{s1_mats_df}{Filtered MATS df for study 1}

\item{s2_mats_df}{Filtered MATS df for study 2}

\item{study_1}{Name for study 1}

\item{study_2}{Name for study 2}

\item{beta_threshold}{Minimum delPSI threshold. AS events below this for either studies are discarded.}

\item{slope}{If TRUE - show slope of the linear fit}
}
\value{
Scatter plot of AS changes betwene those two studies
}
\description{
Generate scatter plot with linear fit statistics between two AS changes
}
