% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_splice_site.R
\name{generate_splice_site_consensus_plot}
\alias{generate_splice_site_consensus_plot}
\title{Main Wrapper to generate Splice Site Analysis Plot}
\usage{
generate_splice_site_consensus_plot(
  pos_events,
  neg_events,
  condition_label_one,
  condition_label_two,
  event_type = "SE",
  padding = 10
)
}
\arguments{
\item{pos_events}{AS events with positive delta PSI - this will be on the top when plots are generated}

\item{neg_events}{AS events with negative delta PSI - this will be on the bot when plots are generated}

\item{event_type}{Event Type - SE, A3SS, A5SS}

\item{condition_label}{labels to be added in the plot}
}
\value{
combined plot with all the plots merged
}
\description{
Main Wrapper to generate Splice Site Analysis Plot
}
