% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_splice_site.R
\name{generate_ss_sequence_logo}
\alias{generate_ss_sequence_logo}
\title{Generate Splice site sequence logo}
\usage{
generate_ss_sequence_logo(object, event_type, padding = 10)
}
\arguments{
\item{object}{splice site information object with ss sequence and ss score}
}
\value{
list of plots with sequence logo
}
\description{
Generate Splice site sequence logo
}
