% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_splice_site.R
\name{generate_ss_sequence_score}
\alias{generate_ss_sequence_score}
\title{Generate Splice Site Strength Density plot}
\usage{
generate_ss_sequence_score(inc_object, exc_object, event_type)
}
\arguments{
\item{inc_object}{splice site information object with ss sequence and ss score for inclusion part (TOP)}

\item{exc_object}{splice site information object with ss sequence and ss score for inclusion part (BOTTOM)}

\item{event_type}{Event Type - SE, A3SS, A5SS}
}
\value{
list of plots with sequence strength distribution
}
\description{
Generate Splice Site Strength Density plot
}
