% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_lollipop_summary_plot.R
\name{generate_strip_lollipop_plot_dev}
\alias{generate_strip_lollipop_plot_dev}
\title{Generate Lollipop Plot with the Strip for specific Splicing Type}
\usage{
generate_strip_lollipop_plot_dev(
  test_result_list,
  rbp_of_interest,
  spl_type,
  max_score_threshold = 0.8,
  color_offset = 0,
  color = "Set1",
  manual_colors = NULL
)
}
\arguments{
\item{test_result_list}{SPARKS analysis result list for SE, A3SS, A5SS}

\item{rbp_of_interest}{RBPs that would be annotated on the lollipop}

\item{spl_type}{Splicing Type - SE, A5SS, A3SS}
}
\value{
Lollipop plot with strip on the bottom for tracks
}
\description{
Generate Lollipop Plot with the Strip for specific Splicing Type
}
