% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_subcluster_heatmaps}
\alias{generate_subcluster_heatmaps}
\title{Generate Subcluster Heatmaps (Per Cell and Aggregated)}
\usage{
generate_subcluster_heatmaps(seurat_obj, genes, out_dir, prefix)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{genes}{Vector of genes to plot}

\item{out_dir}{Output directory path}

\item{prefix}{Prefix for the file names}
}
\description{
Generate Subcluster Heatmaps (Per Cell and Aggregated)
}
