% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARKS_analysis.R
\name{generate_subset_SPARKS_rerun}
\alias{generate_subset_SPARKS_rerun}
\title{Subset SPARKS object for subsets}
\usage{
generate_subset_SPARKS_rerun(
  input_sparks,
  kd_library_all,
  subset_study,
  sample_one_id,
  sample_two_id,
  num_cores = 3,
  overlap_ratio_threshold = 0.3
)
}
\arguments{
\item{input_sparks}{Input SPARKS object}

\item{kd_library_all}{Signature library for all splice type}

\item{subset_study}{Study name for this subset comparison}

\item{sample_one_id}{Column ID for group one.}

\item{sample_two_id}{Column ID for group two.}

\item{num_cores}{Number of cores for SPARKS analysis}

\item{overlap_ratio_threshold}{Threshold to filter out AS events based on their occurrences in the library - default = 30%}
}
\value{
subset_SPARKS_object
}
\description{
Subset SPARKS object for subsets
}
