% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_hotspot_association.R
\name{generate_summary_table_sig_AS_per_HS}
\alias{generate_summary_table_sig_AS_per_HS}
\title{Generate Summary table for each hotspot assocation}
\usage{
generate_summary_table_sig_AS_per_HS(object)
}
\arguments{
\item{object}{SPARKS object}
}
\value{
SPARKS object with the summary df added
}
\description{
Generate Summary table for each hotspot assocation
}
