% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_hotspot_association.R
\name{generate_table_plot_for_num_asso_AS_per_hotspot}
\alias{generate_table_plot_for_num_asso_AS_per_hotspot}
\title{Generate table plot with number of associated AS counts per hotspot}
\usage{
generate_table_plot_for_num_asso_AS_per_hotspot(rbp_hs_sum_df_sorted)
}
\arguments{
\item{rbp_hs_sum_df_sorted}{Hotspot assocation summary heatmap data}
}
\value{
Heatmap
}
\description{
This function generates heatmap plot - row = hotspot, col = splicing type, element = # AS events associated
}
