% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_top_expressed_genes}
\alias{generate_top_expressed_genes}
\title{Top expressed genes table per group}
\usage{
generate_top_expressed_genes(
  seurat_obj,
  group_by_col,
  out_dir,
  prefix,
  top_n = 10L
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{group_by_col}{Character}

\item{out_dir}{Character}

\item{prefix}{Character}

\item{top_n}{Integer}
}
\description{
Writes: TopExpressed_{prefix}_{col}.txt
}
