% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_violin_plots}
\alias{generate_violin_plots}
\title{Save VlnPlots split by condition for each gene}
\usage{
generate_violin_plots(
  seurat_obj,
  genes,
  out_dir,
  prefix,
  group_by_col = "seurat_clusters",
  split_by = "condition"
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{genes}{Character vector}

\item{out_dir}{Character}

\item{prefix}{Character}

\item{group_by_col}{Character. Metadata column for x-axis groups}

\item{split_by}{Character or NULL. Metadata column to split violins}
}
\description{
Save VlnPlots split by condition for each gene
}
