% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_avg_expr}
\alias{get_avg_expr}
\title{Get Average Expression wrapper}
\usage{
get_avg_expr(seurat_obj, layer = "data")
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{layer}{Character. "data" or "counts"}
}
\value{
Matrix of average expression or NULL
}
\description{
Get Average Expression wrapper
}
