% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_junk_pattern}
\alias{get_junk_pattern}
\title{Get regex pattern for junk genes dependent on species}
\usage{
get_junk_pattern(species = "Mouse")
}
\arguments{
\item{species}{Character. "Mouse" or "Human"}
}
\value{
Regex string
}
\description{
Get regex pattern for junk genes dependent on species
}
