% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_valid_groups}
\alias{get_valid_groups}
\title{Get valid groups containing a minimum number of cells}
\usage{
get_valid_groups(meta, col, min_cells = 3L)
}
\arguments{
\item{meta}{Data frame (Seurat metadata)}

\item{col}{Character. Metadata column name}

\item{min_cells}{Integer}
}
\value{
Character vector of valid groups
}
\description{
Get valid groups containing a minimum number of cells
}
