% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_ENCODE_CLIP_intersect_data}
\alias{import_ENCODE_CLIP_intersect_data}
\title{Main import function to import ENCODE CLIP peak count into object}
\usage{
import_ENCODE_CLIP_intersect_data(object, clip_result_file, event_type)
}
\arguments{
\item{object}{the object}

\item{clip_result_file}{CLIP peak overlap file generated from analysis pipeline}

\item{event_type}{event_type}
}
\value{
object with the list of dataframe added
}
\description{
This function reads in the file and add it to the object
}
