% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_exon_annotation}
\alias{import_exon_annotation}
\title{Main import function to import exon annotation into object}
\usage{
import_exon_annotation(object, annotation_file, event_type)
}
\arguments{
\item{object}{the object}

\item{annotation_file}{Exon annotation file (KnownJC/NovelJC/NovelSS) annotated for each AS event}

\item{event_type}{event_type}
}
\value{
object with the list of dataframe added
}
\description{
This function reads in the file and add it to the object
}
