% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_expression_matrix}
\alias{import_expression_matrix}
\title{Main import function to import expression df into object}
\usage{
import_expression_matrix(object, input_exp_file)
}
\arguments{
\item{object}{the object}

\item{input_exp_file}{expression matrix file from the processing pipeline}
}
\value{
object with the expression dataframe added
}
\description{
This function reads in the file and filteres the data and add it to the object
}
