% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_hotspot_maf}
\alias{import_hotspot_maf}
\title{Main import function to import MAF into object}
\usage{
import_hotspot_maf(object, input_maf_file)
}
\arguments{
\item{object}{the object}

\item{input_maf_file}{maf file from analysis pipeline - hotspot should be added in this MAF}
}
\value{
object with the MAF result dataframe added
}
\description{
This function reads in the MAF file and add it to the object
}
