% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_hotspot_mutation_matrix}
\alias{import_hotspot_mutation_matrix}
\title{Main import function to import hotspot matrix into object}
\usage{
import_hotspot_mutation_matrix(object, input_hotspot_matrix_file)
}
\arguments{
\item{object}{the object}

\item{input_hotspot_matrix_file}{hotspot matrix file from the processing pipeline}
}
\value{
object with the expression dataframe added
}
\description{
This function reads in the file and add it to the object
Hotspot matrix is where columns are sample, rows are hotspot mutations,
1 if the sample has mutation in the hotspot, 0 if not
}
