% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_linear_model_data}
\alias{import_linear_model_data}
\title{Main import function to import pval df into object}
\usage{
import_linear_model_data(
  object,
  input_hotspot_lm_pval_file,
  input_hotspot_lm_beta_file,
  spl_event_type,
  neglog_pval_cutoff = 5,
  beta_cutoff = 0.1
)
}
\arguments{
\item{object}{the object}

\item{input_hotspot_lm_pval_file}{pval file from the processing pipeline}

\item{input_hotspot_lm_beta_file}{beta file from the processing pipeline}

\item{spl_event_type}{either SE, A3SS, A5SS}
}
\value{
object with the linear model result dataframe added
}
\description{
This function reads in the file and filteres the data and add it to the object
}
