% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_linear_model_raw_data}
\alias{import_linear_model_raw_data}
\title{Wrapper for cleaning up hotspot pval dataframe}
\usage{
import_linear_model_raw_data(input_hotspot_lm_file, spl_event_type)
}
\arguments{
\item{input_hotspot_lm_file}{hotspot lm result file from the processing pipeline}

\item{spl_event_type}{either SE, A3SS, A5SS}
}
\value{
clean_pval_df - pval df where col/rownames are processed
}
\description{
This function reads in the file and cleans up accordingly
}
