% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_motif_count_data}
\alias{import_motif_count_data}
\title{Main import function to import motif count into object}
\usage{
import_motif_count_data(object, motif_count_dir, spl_event_type)
}
\arguments{
\item{object}{the object}

\item{motif_count_dir}{directory with motif count data generated putative RBP binding motifs}

\item{spl_event_type}{event_type}
}
\value{
object with the list of dataframe added
}
\description{
This function reads in the file and add it to the object
}
