% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_PSI_df}
\alias{import_PSI_df}
\title{Main import function to import PSI df into object}
\usage{
import_PSI_df(object, input_psi_file, spl_event_type)
}
\arguments{
\item{object}{the object}

\item{input_psi_file}{psi file from the processing pipeline}

\item{spl_event_type}{either SE, A3SS, A5SS}
}
\value{
object with the PSI dataframe added
}
\description{
This function reads in the file and cleans up accordingly and add it to the object
}
