% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_rMATS_output.R
\name{import_raw_rMATS_output}
\alias{import_raw_rMATS_output}
\title{Import raw rMATS output}
\usage{
import_raw_rMATS_output(input_dir, spl_type = "SE")
}
\arguments{
\item{input_dir}{directory with rMATS Output}

\item{spl_type}{splicing type (default = SE)}
}
\value{
mats_df
}
\description{
Process raw rMATS output for SPARKS
}
