% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARKS_analysis.R
\name{import_SPARKS_MATS_for_analysis}
\alias{import_SPARKS_MATS_for_analysis}
\title{Import MATS for SPARKS analysis from SPARKS object}
\usage{
import_SPARKS_MATS_for_analysis(
  input_sparks,
  spl_type = "SE",
  count_threshold = 20,
  min_threshold = 20
)
}
\arguments{
\item{input_sparks}{Input SPARKS object}

\item{spl_type}{Optional splicing type in c("SE", "A3SS", "A5SS"), default = "SE"}

\item{count_threshold}{Average read count for AS events. Events with read counts below this number in average is removed}

\item{min_threshold}{Minimum read count for AS events. Events with read counts below this number in any sample is removed}
}
\value{
filtered_MATS_df
}
\description{
This function reads MATS file generated from the pipeline, keep known events,
and filter out events with the read counts below the count_threshold in any
samples.
}
\details{
It also keeps only necessary information for the AS events defining junctions.
}
