% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_pipeline_config}
\alias{load_pipeline_config}
\title{Load and merge a base config and an optional override config}
\usage{
load_pipeline_config(base_config_path, override_config_path = NULL)
}
\arguments{
\item{base_config_path}{Character. Path to base YAML (template)}

\item{override_config_path}{Character. Path to override YAML (your file)}
}
\value{
Merged named list
}
\description{
Load and merge a base config and an optional override config
}
