% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_dir}
\alias{make_dir}
\title{Create directory recursively if it doesn't exist}
\usage{
make_dir(path)
}
\arguments{
\item{path}{Character. Directory path}
}
\value{
Invisible path
}
\description{
Create directory recursively if it doesn't exist
}
