% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{normalize}
\usage{
normalize(MATRIX)
}
\arguments{
\item{hi}{ya}
}
\value{
yo
}
\description{
normalize each column of a matrix into unit vector. if the column is filled with zeros, it is left unchanged, since calculating the l2 norm of that column and dividing the column by that norm would lead to NA (diviing by zero)
}
\examples{
hi

}
