% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_CLIP_data.R
\name{perform_CLIP_enrichment_test_direction}
\alias{perform_CLIP_enrichment_test_direction}
\title{Perform CLIP Enrichment Test for All regions individually for directionality}
\usage{
perform_CLIP_enrichment_test_direction(
  object,
  pos_events,
  neg_events,
  event_type = "SE"
)
}
\arguments{
\item{object}{SPARKS object}

\item{pos_events}{AS events of interest in one group}

\item{event_type}{either SE, A3SS, A5SS}

\item{event_anno}{AS events of interest in the other group}
}
\value{
Result df in 5 columns - "pval", "odds_ratio", "peak_count", "RBP", "Region"
}
\description{
This function performs test for enrichment in CLIP peaks in each region
relative to splicing invidiaully.
This function is designed to test directional enrichment, given positive and negative events
}
