% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_CLIP_data.R
\name{perform_CLIP_enrichment_test_regional}
\alias{perform_CLIP_enrichment_test_regional}
\title{Perform CLIP Enrichment Test for All regions individually}
\usage{
perform_CLIP_enrichment_test_regional(
  object,
  as_events,
  event_type = "SE",
  event_anno = c(),
  all_events = c()
)
}
\arguments{
\item{object}{SPARKS object}

\item{as_events}{AS events of interest}

\item{event_type}{either SE, A3SS, A5SS}

\item{event_anno}{Junction information - KnownJC, NovelJC, NovelSS - option to limit}

\item{all_events}{Background + Foreground AS events to limit the scope of the analysis}
}
\value{
Result df in 5 columns - "pval", "odds_ratio", "peak_count", "RBP", "Region"
}
\description{
This function performs test for enrichment in CLIP peaks in each region
relative to splicing invidiaully.
The function includes "event_anno" to define AS event junction type, whether
known AS events or novel combination of junctions or novel splice sites.
Also this function supports background setting via all_events.
}
