% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARKS_analysis.R
\name{perform_SPARKS_analysis_for_all_splice_types}
\alias{perform_SPARKS_analysis_for_all_splice_types}
\title{Wrapper for running SPARKS for all splice types}
\usage{
perform_SPARKS_analysis_for_all_splice_types(
  input_sparks,
  kd_library_all,
  test_study = "",
  subset_group_1 = c(),
  subset_group_2 = c()
)
}
\arguments{
\item{input_sparks}{Input SPARKS object}

\item{kd_library_all}{Signature library for all splice types}

\item{test_study}{Study name}
}
\value{
SPARKS analysis results for all splice types bound in a list
}
\description{
Wrapper for running SPARKS for all splice types
}
