% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARKS_analysis.R
\name{perform_SPARKS_analysis_with_overlap_filter}
\alias{perform_SPARKS_analysis_with_overlap_filter}
\title{Run SPARKS analysis}
\usage{
perform_SPARKS_analysis_with_overlap_filter(
  study_mats,
  kd_library,
  study,
  num_cores = 3,
  overlap_ratio_threshold = 0.3,
  library_list = c(),
  event_count_df = data.frame(),
  library_size = 675
)
}
\arguments{
\item{study_mats}{filtered MATS dataframe}

\item{kd_library}{Signature library}

\item{study}{Study name - this will be stored in "S2" column in the output}

\item{num_cores}{Number of cores for parallel run - default = 3}

\item{overlap_ratio_threshold}{Threshold to filter out AS events based on their occurrences in the library - default = 30%}

\item{library_list}{If given, SPARKS analysis will be performed in only this signatures}

\item{event_count_df}{If given, SPARKS will use this to enumerate events to filter out, rather than iterating the library}

\item{library_size}{Library size - currently this is 675. This is built in case for future cases where the signature library is expanded. Do not change this for now.}
}
\value{
SPARKS_analysis_df
}
\description{
Run SPARKS analysis
}
