% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_hotspot_association.R
\name{query_associated_AS_events}
\alias{query_associated_AS_events}
\title{Query list of AS events associated for given Mutation event}
\usage{
query_associated_AS_events(object, mutation_event)
}
\arguments{
\item{object}{SPARKS object}

\item{mutation_event}{Mutation event}
}
\value{
list of AS events associated for given Mutation event
}
\description{
This function generates heatmap plot - x = hotspot mutation status, y = PSI
}
