% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_CLIP_data.R
\name{query_clip_peak_df_for_AS_events}
\alias{query_clip_peak_df_for_AS_events}
\title{Query CLIP peak df for given AS events}
\usage{
query_clip_peak_df_for_AS_events(object, as_events, event_type = "SE")
}
\arguments{
\item{object}{SPARKS object}

\item{as_events}{AS events to get query for}
}
\value{
CLIP peak dataframe - columns are region, rows are AS events, elements are whether it has peak at that position or not
}
\description{
Query CLIP peak df for given AS events
}
