% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_CLIP_data.R
\name{query_clip_peak_df_for_RBP}
\alias{query_clip_peak_df_for_RBP}
\title{Query CLIP peak df for given RBP}
\usage{
query_clip_peak_df_for_RBP(object, target, event_type = "SE")
}
\arguments{
\item{object}{SPARKS object}

\item{target}{Specific RBPs to query CLIP peaks for}
}
\value{
CLIP peak dataframe - columns are region, rows are AS events, elements are whether it has peak at that position or not
}
\description{
Query CLIP peak df for given RBP
}
