% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_expression_data.R
\name{query_expression_data_for_gene_set}
\alias{query_expression_data_for_gene_set}
\title{Query expression data from SPARKS object}
\usage{
query_expression_data_for_gene_set(object, gene_list)
}
\arguments{
\item{object}{SPARKS object}

\item{gene_list}{list of genes you are interested in}
}
\value{
Dataframe of expression values for the select genes
}
\description{
Query expression data from SPARKS object
}
