% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_accessory_functions.R
\name{query_PSI_value_for_AS_event}
\alias{query_PSI_value_for_AS_event}
\title{Query PSI value for AS event}
\usage{
query_PSI_value_for_AS_event(object, AS_event)
}
\arguments{
\item{object}{SPARKS object}

\item{event_type}{AS event}
}
\value{
psi values for the AS event
}
\description{
Query PSI value for AS event
}
