% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_splice_site.R
\name{query_splice_site_sequence_for_AS_events}
\alias{query_splice_site_sequence_for_AS_events}
\title{Query Splice site sequence for given AS events}
\usage{
query_splice_site_sequence_for_AS_events(as_events, padding = 10)
}
\arguments{
\item{as_events}{AS events to get query for}
}
\value{
splice site information object sith ss sequence and ss score
}
\description{
Query Splice site sequence for given AS events
}
