% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPARKS_analysis.R
\name{rewrite_event_coordinates}
\alias{rewrite_event_coordinates}
\title{Rewrite event coordinates}
\usage{
rewrite_event_coordinates(x)
}
\arguments{
\item{x}{AS event}
}
\value{
filtered_AS_event
}
\description{
This function removes unnecessary info in AS event coordinates.
}
