% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvd_spca.R
\name{rsvd_spca}
\alias{rsvd_spca}
\title{rsvd_spca}
\usage{
rsvd_spca(dat, R, lambda, penalty = c("soft", "scad"), ridge = 1e-06,
  maxiter, scadgamma, inits = c("SVD", "oracle", "multistart"),
  nrstart = NULL, oracle = NULL)
}
\arguments{
\item{hi}{ya}
}
\value{
yo
}
\description{
implementation of rsvd_spca method by Shen and Huang. only soft-thresholding for now (december 2018)
}
\examples{
hi

}
