% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{run_analysis_unit}
\alias{run_analysis_unit}
\title{Run all analyses for a Seurat object (main or subset)}
\usage{
run_analysis_unit(
  seurat_obj,
  display_name,
  groupings,
  genes_list,
  base_dir,
  suffix,
  deg_color,
  cfg
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{display_name}{Character}

\item{groupings}{Character vector. Metadata columns to loop over}

\item{genes_list}{Character vector}

\item{base_dir}{Character}

\item{suffix}{Character}

\item{deg_color}{Character}

\item{cfg}{Named list}
}
\description{
Run all analyses for a Seurat object (main or subset)
}
