% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_analysis.R
\name{run_escape_enrichment}
\alias{run_escape_enrichment}
\title{Run escape pathway enrichment analysis}
\usage{
run_escape_enrichment(
  seurat_obj,
  species = "Mouse",
  library = "H",
  method = "ssGSEA",
  min_size = 5
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{species}{Character. "Mouse" or "Human"}

\item{library}{Character. e.g., "H" for Hallmark, "C5" for GO}

\item{method}{Character. e.g., "ssGSEA", "UCell", "AUCell"}

\item{min_size}{Integer. Min genes in a pathway.}
}
\value{
Seurat object with escape assay added
}
\description{
Run escape pathway enrichment analysis
}
