% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_core.R
\name{run_grouping_analysis}
\alias{run_grouping_analysis}
\title{Run all analyses for one grouping column}
\usage{
run_grouping_analysis(
  seurat_obj,
  group_col,
  file_prefix,
  genes_list,
  base_dir,
  suffix,
  deg_color,
  cfg
)
}
\arguments{
\item{seurat_obj}{Seurat object. Must already have PrepSCTFindMarkers applied.}

\item{group_col}{Character}

\item{file_prefix}{Character}

\item{genes_list}{Character vector}

\item{base_dir}{Character}

\item{suffix}{Character}

\item{deg_color}{Character}

\item{cfg}{Named list}
}
\description{
Run all analyses for one grouping column
}
