% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline_core.R
\name{run_pipeline}
\alias{run_pipeline}
\title{Run the complete scRNA-seq pipeline}
\usage{
run_pipeline(
  base_config_path,
  override_config_path = NULL,
  sample_metadata = NULL
)
}
\arguments{
\item{base_config_path}{Path to the template YAML}

\item{override_config_path}{Path to your custom override YAML}

\item{sample_metadata}{Data frame}
}
\description{
Run the complete scRNA-seq pipeline
}
