% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{run_scproportion_test}
\alias{run_scproportion_test}
\title{scProportionTest}
\usage{
run_scproportion_test(seurat_obj, grouping_col, output_dir, file_prefix)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{grouping_col}{Character}

\item{output_dir}{Character}

\item{file_prefix}{Character}
}
\description{
scProportionTest
}
