% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{run_singler_annotation}
\alias{run_singler_annotation}
\title{Annotate cells with SingleR (primary + secondary references)}
\usage{
run_singler_annotation(
  seurat_obj,
  species_target,
  ref_celldex1,
  ref_celldex2,
  singler_cfg
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{species_target}{species for celldex db}

\item{ref_celldex1}{celldex reference object 1}

\item{ref_celldex2}{celldex reference object 2}

\item{singler_cfg}{Named list from config}
}
\value{
Seurat object with annotation columns added
}
\description{
Annotate cells with SingleR (primary + secondary references)
}
