% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{save_average_expression}
\alias{save_average_expression}
\title{Compute and save average expression per group × condition}
\usage{
save_average_expression(
  seurat_obj,
  output_dir,
  file_prefix,
  group_by_col = "seurat_clusters",
  layer = "data",
  table_sep = "\\t",
  table_quote = FALSE,
  table_row_names = FALSE
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{output_dir}{Character}

\item{file_prefix}{Character}

\item{group_by_col}{Character}

\item{layer}{Character. "data" or "counts"}

\item{table_sep}{Character}

\item{table_quote}{Logical}

\item{table_row_names}{Logical}
}
\description{
Writes: AvgExpr_{layer}_{prefix}_by_{col}.txt
}
