% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_png}
\alias{save_png}
\title{Save a plot to PNG using ggsave or base R for pheatmap/gtable}
\usage{
save_png(p, filename, width = 8, height = 6, dpi = 300)
}
\arguments{
\item{p}{Plot object (ggplot or gtable)}

\item{filename}{Character. Output file path}

\item{width}{Numeric}

\item{height}{Numeric}

\item{dpi}{Numeric}
}
\description{
Save a plot to PNG using ggsave or base R for pheatmap/gtable
}
