% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleData.R
\name{scaleData}
\alias{scaleData}
\title{scaleData}
\usage{
scaleData(X, value = 0)
}
\arguments{
\item{hi}{}
}
\value{
hi
}
\description{
idea/code 100 percent from Niek. Thanks sir. Standardizes the columns using the "biased sd" ####
}
\examples{
hi

}
