% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsvd_spca_function.R
\name{softVec}
\alias{softVec}
\title{rsvd_spca}
\usage{
softVec(alpha, lambda)
}
\arguments{
\item{hi}{ya}
}
\value{
yo
}
\description{
implementation of rsvd_spca method by Shen and Huang. only soft-thresholding for now (december 2018)
}
\examples{
hi

}
