% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca_adj.R
\name{spca_adj}
\alias{spca_adj}
\title{spca_adj}
\usage{
spca_adj(x, K, para, type = c("predictor", "Gram"),
  sparse = c("penalty", "varnum"), use.corr = FALSE, lambda = 1e-06,
  max.iter = 200, trace = FALSE, eps.conv = 0.001, inits = c("SVD",
  "oracle", "multistart"), nrstart = NULL, oracle = NULL)
}
\arguments{
\item{hi}{}
}
\value{
hi
}
\description{
adjusted the spca function within the package elasticnet. uses most of the code of the spca() function in elasticnet. also uses the solvebeta function from elasticnet package.
}
\examples{
hi

}
