% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_hotspot_association.R
\name{subset_genes_of_interest_from_summary}
\alias{subset_genes_of_interest_from_summary}
\title{Subset the summary}
\usage{
subset_genes_of_interest_from_summary(object, genes_of_interest)
}
\arguments{
\item{object}{SPARKS object}

\item{genes_of_interest}{Genes that you want subset for in terms of hotspot}
}
\value{
Subsetted summary dataframe for downstream functions
}
\description{
Subset the summary
}
