% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{update_sample_chart}
\alias{update_sample_chart}
\title{Internal update function to update the sample chart}
\usage{
update_sample_chart(object, data_name, samples)
}
\arguments{
\item{object}{the object}

\item{data_name}{column name to annotate the data}

\item{samples}{samples that have certain features - 0/1 will be added accordingly}
}
\value{
updated object with additional annotation
}
\description{
This function updates the sample chart to add extra annotations arising from analysis
}
