% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_table}
\alias{write_table}
\title{Write table helper}
\usage{
write_table(x, file, sep = "\\t", quote = FALSE, row_names = FALSE)
}
\arguments{
\item{x}{Data frame or matrix}

\item{file}{Character. Output file path}

\item{sep}{Character. Delimiter}

\item{quote}{Logical}

\item{row_names}{Logical}
}
\description{
Write table helper
}
