% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wsparsedata.R
\name{Wsparsedata}
\alias{Wsparsedata}
\title{Wsparsedata}
\usage{
Wsparsedata(n, p, k, n_zeros = NULL, Winit = NULL, empirical = FALSE,
  seed, VAFx)
}
\arguments{
\item{hi}{}
}
\value{
hi
}
\description{
generates data from a sparse weights structure
}
\examples{
hi

}
