% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{.load_counts}
\alias{.load_counts}
\title{Load count matrix from Alevin or Cell Ranger output}
\usage{
.load_counts(folder_id, data_path)
}
\arguments{
\item{folder_id}{Character}

\item{data_path}{Character. Root data directory}
}
\value{
Raw count matrix
}
\description{
Load count matrix from Alevin or Cell Ranger output
}
\keyword{internal}
