% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{process_single_sample}
\alias{process_single_sample}
\title{Process a single sample: load, filter, doublet removal}
\usage{
process_single_sample(
  folder_id,
  protocol,
  file_prefix,
  qc_dir,
  data_path,
  gene_removal_pattern,
  mt_pattern,
  genes_to_remove,
  min_features,
  max_counts,
  max_mt_percent,
  min_cells,
  cfg
)
}
\arguments{
\item{folder_id}{Character}

\item{protocol}{Character. Condition label (e.g. "WT")}

\item{file_prefix}{Character}

\item{qc_dir}{Character}

\item{data_path}{Character}

\item{gene_removal_pattern}{Character regex}

\item{mt_pattern}{Character regex}

\item{genes_to_remove}{Character vector}

\item{min_features}{Integer}

\item{max_counts}{Integer}

\item{max_mt_percent}{Numeric}

\item{min_cells}{Integer}

\item{cfg}{Named list. Full pipeline config}
}
\value{
Seurat object or NULL
}
\description{
Process a single sample: load, filter, doublet removal
}
