% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{run_proportion_analysis}
\alias{run_proportion_analysis}
\title{Compute cell type proportion table and chi-squared test}
\usage{
run_proportion_analysis(seurat_obj, grouping_col, output_dir, file_prefix)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{grouping_col}{Character}

\item{output_dir}{Character}

\item{file_prefix}{Character}
}
\description{
Compute cell type proportion table and chi-squared test
}
