% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{save_pseudobulk_counts}
\alias{save_pseudobulk_counts}
\title{Compute and save pseudobulk counts per sample x group}
\usage{
save_pseudobulk_counts(
  seurat_obj,
  output_dir,
  file_prefix,
  group_by_col = "seurat_clusters",
  table_sep = "\\t",
  min_replicates = 3L
)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{output_dir}{Character}

\item{file_prefix}{Character}

\item{group_by_col}{Character}

\item{table_sep}{Character}

\item{min_replicates}{Integer. Minimum unique samples required for pseudobulking}
}
\description{
Writes: PseudobulkCounts_{prefix}_by_{col}.txt
}
